Make all const variables var (#34)
This makes it so that the consumer doesn't have to care about ES6 support when using this library as standalone. In my case that is UglifyJS in my build pipeline which does not support ES6. I could probably make it so that I transpile before minifying the code, but there could be other people who may run into this issue as well. And also, using `var` seem to match the library style better anyhow.

An alternative would be to make a more elaborate build pipeline for this library to supply a dist library, which seem to have been the case previously, but was removed.